requirements.txt
# LangChain and LangGraph dependencies
langchain>=1.0.0
langchain-ollama>=0.1.0
langchain-core>=0.3.0
langgraph>=0.2.0

# PostgreSQL adapter
psycopg2-binary>=2.9.9

# MCP SDK
mcp>=0.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.9.0
